用分布式方案替代router-server

+++ talk-common/src/main/java/com/blt/talk/common/code/IMHeader.java
@@ -211,7 +211,7 @@ public class IMHeader {
      * @return the reserved
      * @since  1.0
      */
-    public int getReserved() {
+    public short getReserved() {
         return reserved;
     }
